+++ R/Run_Mod.R
@@ -48,14 +48,16 @@
 #'   - `nParameters`: The amount of parameters estimated in the model.
 #'   - `nIterations`: Number of iteration specified.
 #'   - `nBurnin`: Number of burn ins.
-#'   - `nThin`: Thinning number (niter/nmc)
+#'   - `nThin`: Thinning number (niter/nmc).
 #'   - `priors`: A [list] that summarizes the input priors, including:
 #'     - `mu_hyp_param`
 #'     - `prec_hyp_param`
 #'     - `omega_param`
 #'     - `wishdf`
 #'     - `prec_logy_hyp_param`
-#'     - An optional `"jags.post"` attribute, included when argument
+#'   - `fitted_residuals`: A [data.frame] containing fitted and residual values
+#'   for all observations.
+#'   - An optional `"jags.post"` attribute, included when argument
 #'   `with_post` = TRUE.
 #' @inheritDotParams prep_priors
 #' @export
@@ -192,11 +194,18 @@ run_mod <- function(data,
   jags_out <- jags_out |>
     structure("priors" = attributes(priorspec)$used_priors)
   
+  # Calculating fitted and residuals
+  # Renaming columns using attributes from as_case_data
+  fit_res <- calc_fit_mod(modeled_dat = jags_out,
+                          original_data = dl_sub)
+  jags_out <- jags_out |>
+    structure(fitted_residuals = fit_res)
+
   # Conditionally adding jags.post
   if (with_post) {
     jags_out <- jags_out |>
       structure(jags.post = jags_post_final)
-  } 
+  }
   jags_out <- jags_out |>
     structure(class = union("sr_model", class(jags_out)))
   jags_out