+++ tests/testthat/test-run_mod.R
@@ -1,6 +1,8 @@
 test_that(
   desc = "results are consistent with simulated data",
   code = {
+    testthat::announce_snapshot_file("sim-strat-curve-params.csv")
+    testthat::announce_snapshot_file("sim-strat-fitted_residuals.csv")
     skip_on_os(c("windows", "linux"))
     withr::local_seed(1)
     strat1 <- serocalculator::typhoid_curves_nostrat_100 |>
@@ -31,11 +33,14 @@ test_that(
         
         results |>
           attributes() |>
-          rlist::list.remove("row.names") |>
+          rlist::list.remove(c("row.names", "fitted_residuals")) |>
           expect_snapshot_value(style = "deparse")
         
         results |>
           ssdtools:::expect_snapshot_data("sim-strat-curve-params")
+
+        attributes(results)$fitted_residuals |>
+          ssdtools:::expect_snapshot_data("sim-strat-fitted_residuals")
         
       }
     )
@@ -45,6 +50,8 @@ test_that(
 test_that(
   desc = "results are consistent with SEES data",
   code = {
+    testthat::announce_snapshot_file("strat-curve-params.csv")
+    testthat::announce_snapshot_file("strat-fitted_residuals.csv")
     skip_on_os(c("windows", "linux"))
     withr::local_seed(1)
     dataset <- serodynamics::nepal_sees 
@@ -63,17 +70,22 @@ test_that(
 
     results |>
       attributes() |>
-      rlist::list.remove("row.names") |>
+      rlist::list.remove(c("row.names", "fitted_residuals")) |>
       expect_snapshot_value(style = "deparse")
 
     results |>
       ssdtools:::expect_snapshot_data("strat-curve-params")
+
+    attributes(results)$fitted_residuals |>
+      ssdtools:::expect_snapshot_data("strat-fitted_residuals")
   }
 )
 
 test_that(
   desc = "results are consistent with unstratified SEES data",
   code = {
+    announce_snapshot_file("nostrat-curve-params.csv")
+    announce_snapshot_file("nostrat-fitted_residuals.csv")
     skip_on_os(c("windows", "linux"))
     withr::local_seed(1)
     dataset <- serodynamics::nepal_sees 
@@ -92,11 +104,14 @@ test_that(
 
     results |>
       attributes() |>
-      rlist::list.remove("row.names") |>
+      rlist::list.remove(c("row.names", "fitted_residuals")) |>
       expect_snapshot_value(style = "deparse")
 
     results |>
       ssdtools:::expect_snapshot_data("nostrat-curve-params")
+
+    attributes(results)$fitted_residuals |>
+      ssdtools:::expect_snapshot_data("nostrat-fitted_residuals")
   }
 )
 
@@ -104,6 +119,7 @@ test_that(
   desc = "results are consistent with unstratified SEES data with jags.post
   included",
   code = {
+    announce_snapshot_file("nostrat-curve-params-withpost.csv")
     skip_on_os(c("windows", "linux"))
     withr::local_seed(1)
     dataset <- serodynamics::nepal_sees 
@@ -123,7 +139,7 @@ test_that(
     
     results |>
       attributes() |>
-      rlist::list.remove(c("row.names", "jags.post")) |>
+      rlist::list.remove(c("row.names", "jags.post", "fitted_residuals")) |>
       expect_snapshot_value(style = "serialize")
     
     results |>
@@ -135,6 +151,7 @@ test_that(
   desc = "results are consistent with unstratified SEES data with modified 
   priors",
   code = {
+    announce_snapshot_file("nostrat-curve-params-specpriors.csv")
     skip_on_os(c("windows", "linux"))
     withr::local_seed(1)
     dataset <- serodynamics::nepal_sees 
@@ -158,7 +175,7 @@ test_that(
     
     results |>
       attributes() |>
-      rlist::list.remove(c("row.names")) |>
+      rlist::list.remove(c("row.names", "fitted_residuals")) |>
       expect_snapshot_value(style = "serialize")
     
     results |>